requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

